move build to maven; upgrade to geotrellis 1.0.0; fix unit tests; development cycle 0.11.0
