--- gdal/src/main/scala/geotrellis/gdal/io/hadoop/GdalHadoopSparkContextMethods.scala
@@ -1,78 +1,0 @@
-/*
- * Copyright (c) 2014 Azavea.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package geotrellis.gdal.io.hadoop
-
-import geotrellis.raster._
-import geotrellis.spark.io.hadoop._
-import geotrellis.spark.io.hadoop.HadoopSparkContextMethods
-import geotrellis.spark.TemporalProjectedExtent
-import geotrellis.gdal.io.hadoop.GdalInputFormat._
-
-import org.apache.hadoop.fs.Path
-import org.apache.spark.rdd.RDD
-
-
-trait GdalHadoopSparkContextMethods extends HadoopSparkContextMethods {
-  def gdalRDD(path: Path): RDD[(GdalRasterInfo, Tile)] = {
-    val updatedConf = sc.hadoopConfiguration.withInputDirectory(path)
-
-    sc.newAPIHadoopRDD(
-      updatedConf,
-      classOf[GdalInputFormat],
-      classOf[GdalRasterInfo],
-      classOf[Tile]
-    )
-  }
-
-  def netCdfRDD(
-    path: Path,
-    inputFormat: NetCdfInputFormat = DefaultNetCdfInputFormat): RDD[(TemporalProjectedExtent, Tile)] = {
-    val makeTime = (info: GdalRasterInfo) =>
-    info.file.meta.find {
-      case(key, value) => key.toLowerCase == inputFormat.baseDateMetaDataKey.toLowerCase
-    }.map(_._2) match {
-      case Some(baseString) => {
-
-        val (typ, base) = NetCdfInputFormat.readTypeAndDate(
-          baseString,
-          inputFormat.dateTimeFormat,
-          inputFormat.yearOffset,
-          inputFormat.monthOffset,
-          inputFormat.dayOffset
-        )
-
-        info.bandMeta.find {
-          case(key, value) => key.toLowerCase == "netcdf_dim_time"
-        }.map(_._2) match {
-          case Some(s) => NetCdfInputFormat.incrementDate(typ, s.toDouble, base)
-          case _ => base
-        }
-      }
-      case None => throw new IllegalArgumentException("Can't find base date!")
-    }
-
-    gdalRDD(path)
-      .map { case (info, tile) =>
-        val band = TemporalProjectedExtent(
-          extent = info.file.rasterExtent.extent,
-          crs = info.file.crs,
-          time = makeTime(info)
-        )
-        band -> tile
-    }
-  }
-}
