--- gdal/src/main/scala/geotrellis/gdal/io/hadoop/package.scala
@@ -1,9 +1,0 @@
-package geotrellis.gdal.io
-
-import geotrellis.gdal.io.hadoop.GdalHadoopSparkContextMethods
-import org.apache.spark.SparkContext
-
-package object hadoop {
-  implicit class withGdalHadoopSparkContextMethods(val sc: SparkContext)
-      extends GdalHadoopSparkContextMethods
-}
