--- project/plugins.sbt
@@ -1,10 +1,0 @@
-resolvers ++= Seq(
-  Classpaths.sbtPluginReleases,
-  Opts.resolver.sonatypeReleases
-)
-
-addSbtPlugin("net.virtual-void" % "sbt-dependency-graph" % "0.7.5")
-
-addSbtPlugin("com.eed3si9n" % "sbt-assembly" % "0.13.0")
-
-addSbtPlugin("me.lessis" % "bintray-sbt" % "0.3.0")
