+++ src/main/scala/geotrellis/gdal/io/hadoop/package.scala
@@ -1,0 +1,8 @@
+package geotrellis.gdal.io
+
+import org.apache.spark.SparkContext
+
+package object hadoop {
+  implicit class withGdalHadoopSparkContextMethods(val sc: SparkContext)
+      extends GdalHadoopSparkContextMethods
+}
